Refactor login page to handle form with controlled inputs
